changing tag names and refactoring code
